OM4 with CORE2 dynamic forcing option
I created a CORE2_data dummy experiment to contain the dynamic forcing
files.

I moved the saltrestore since it needs to change depending on whether
the dynamic or climatological forcing is used.

I created a helper experiment for the baseline experiment, so that we
don't have to repeat the runtime and curator tags. Then both the
baseline CORE and baseline CORE2 experiments inherit from it.