Fix the atmospheric chemistry input file
- This fix changes an  input file per Larry:
One of the atmospheric chemistry input files, needed for the full-chem "j" and "k" configurations, was accidentally changed. I restored the correct file, but to a new location. If your runs are crashing with an error from READ_SIM_CHM, you can fix it by making the following two simple updates:

1) Replace the input file:

$CPERM/Fabien.Paulot/awg/INPUT/tropchem/sim.AM4_isop_e90.dat

with:

$CPERM/Larry.Horowitz/awg/INPUT/tropchem/sim.AM4_isop_e90_so2h.dat

2) Change the sim_data_filename setting in the tropchem_driver_nml namelist to:

      sim_data_filename = 'sim.AM4_isop_e90_so2h.dat'

That should fix the problem.